Add graphQL files

+++ pkg/domain/types/const.go
@@ -5,5 +5,5 @@ const (
 
 	DefaultMaxSequences = 32
 
-	DefaultAttributeTTL int64 = 3600 * 24 // 1 day
+	DefaultAttributeTTL int = 3600 * 24 // 1 day
 )
